+++ src/test/java/io/parallec/core/main/http/request/template/ParallelClientVarReplacementHostSpecificHeaderTest.java
@@ -28,6 +28,13 @@ public class ParallelClientVarReplacementHostSpecificHeaderTest extends
         pc = new ParallelClient();
         serverThread = new HttpServerThread();
         serverThread.start();
+        
+        //add sleep to make sure the server starts first
+        try {
+            Thread.sleep(500L);
+        } catch (Exception e) {
+            ;
+        }        
     }
 
     @AfterClass