+++ src/test/java/io/parallec/core/main/ping/package-info.java
@@ -1,7 +1,4 @@
 /**
- * 
- */
-/**
  * @author Yuanteng (Jeff) Pei
  *
  */