+++ CHANGELOG.md
@@ -3,7 +3,7 @@ Parallec Change Log
 
 ## Version 0.9.1
 
-_2015-11-03_
+_2015-11-10_
 
 * Chnage: maven pom: maven-compiler-plugin version to 1.7.
 * Fix: aggregateResultMap change to concurrentHashMap: Issue #11